errors.rs wip
